// my-video/src/GeneratedProductView.tsx
export const emptyProductData = {
    title: "", category: "", price: "", quantity: "", unit: "kg", description: "",
    country: "", state: "", minOrderQuantity: "", maturity: "", image: "", image2: "", image3: "",
    packaging: "", packagingSize: "", shippingUnit: "", containerSize: "", companyName: "",
    contactEmail: "", countryCode: "", phoneNumber: "", certifications: "",
    incoterm: "A definir con el comprador", saleMethod: "standard", supplyCapacity: "",
    supplyCapacityUnit: "toneladas", supplyCapacityPeriod: "mes", currency: "US$",
  };